requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
faiss-cpu==1.13.2
fastapi==0.128.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
llvmlite==0.46.0
MarkupSafe==3.0.3
modelscope==1.33.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai==2.14.0
openai-whisper==20250625
packaging==25.0
Pillow==10.0.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.21
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
yarl==1.22.0
